Move and remove some tests from expressions.spec.ts
